requirements.txt
langchain==0.1.2
langchain_google_genai
langchain_community
huggingface_hub
huggingface-hub
InstructorEmbedding
sentence-transformers==2.2.2
python-dotenv==1.0.0
tiktoken==0.4.0
faiss-cpu==1.7.4
altair==5.3.0
annotated-types==0.6.0
appdirs==1.4.4
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dateutils==0.6.12
decorator==5.1.1
duckdb==0.10.0
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.51.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
gw_dsl_parser==0.1.45a6
idna==3.6
imageio==2.34.0
ipython==8.23.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.10
kanaries_track==0.0.4
kiwisolver==1.4.5
lazy_loader==0.4
lime==0.2.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
networkx==3.3
numpy==1.26.4
packaging
pandas==2.2.1
parso==0.8.4
patsy==0.5.6
pillow==10.3.0
plotly==5.20.0
plotly-express==0.4.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.2
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pygwalker==0.4.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.0
segment-analytics-python==2.2.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.29
sqlglot==23.8.0
stack-data==0.6.3
statsmodels==0.14.1
streamlit
streamlit-option-menu==0.3.12
tenacity==8.2.3
threadpoolctl==3.4.0
tifffile==2024.2.12
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wasmtime==12.0.0
watchdog==4.0.0
wcwidth==0.2.13
widgetsnbextension==4.0.10